+++ posthog/api/__init__.py
@@ -105,14 +105,14 @@ from ee.clickhouse.views.events import ClickhouseEventsViewSet, LegacyClickhouse
 from ee.clickhouse.views.experiments import ClickhouseExperimentsViewSet
 from ee.clickhouse.views.groups import ClickhouseGroupsTypesView, ClickhouseGroupsView
 from ee.clickhouse.views.insights import ClickhouseInsightsViewSet
-from ee.clickhouse.views.person import ClickhousePersonViewSet, LegacyClickhousePersonViewSet
 from ee.clickhouse.views.session_recordings import ClickhouseSessionRecordingViewSet
 from posthog.api.action import ActionViewSet
 from posthog.api.insight import InsightViewSet
+from posthog.api.person import LegacyPersonViewSet, PersonViewSet
 
 # Legacy endpoints CH (to be removed eventually)
 router.register(r"event", LegacyClickhouseEventsViewSet, basename="event")  # Should be completely unused now
-router.register(r"person", LegacyClickhousePersonViewSet, basename="person")
+router.register(r"person", LegacyPersonViewSet, basename="person")
 router.register(r"element", LegacyClickhouseElementViewSet, basename="element")
 router.register(r"cohort", LegacyClickhouseCohortViewSet, basename="cohort")
 # Nested endpoints CH
@@ -121,7 +121,7 @@ projects_router.register(r"events", ClickhouseEventsViewSet, "project_events", [
 projects_router.register(r"groups", ClickhouseGroupsView, "project_groups", ["team_id"])
 projects_router.register(r"groups_types", ClickhouseGroupsTypesView, "project_groups_types", ["team_id"])
 projects_router.register(r"insights", InsightViewSet, "project_insights", ["team_id"])
-projects_router.register(r"persons", ClickhousePersonViewSet, "project_persons", ["team_id"])
+projects_router.register(r"persons", PersonViewSet, "project_persons", ["team_id"])
 projects_router.register(r"elements", ClickhouseElementViewSet, "project_elements", ["team_id"])
 projects_router.register(r"cohorts", ClickhouseCohortViewSet, "project_cohorts", ["team_id"])
 projects_router.register(r"experiments", ClickhouseExperimentsViewSet, "project_experiments", ["team_id"])
